% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth}
\alias{auth}
\title{Authentication method}
\usage{
auth(req, client, auth_type = c("basic", "oauth"))
}
\arguments{
\item{req}{A httr2 request}

\item{client}{list: List with client id and secret. Defaults to the
`MET_FROST_ID` and `MET_FROST_SECRET` environment variables.}

\item{auth_type}{character: Authentication method, either 'basic' or 'oauth'.}
}
\description{
Select authentication method.
}
\keyword{internal}
