% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query.R
\name{create_query}
\alias{create_query}
\title{Create query}
\usage{
create_query(
  endpoint,
  client,
  format,
  version,
  calculation_method = NULL,
  categories = NULL,
  cf_standard_names = NULL,
  cf_cell_methods = NULL,
  cf_units = NULL,
  cf_versions = NULL,
  code_tables = NULL,
  country = NULL,
  county = NULL,
  descriptions = NULL,
  durations = NULL,
  elements = NULL,
  exposure_categories = NULL,
  external_ids = NULL,
  fields = NULL,
  frequencies = NULL,
  geometry = NULL,
  ids = NULL,
  icao_code = NULL,
  include_extra = NULL,
  lang = NULL,
  levels = NULL,
  level_units = NULL,
  level_types = NULL,
  limit = NULL,
  location = NULL,
  max_age = NULL,
  municipality = NULL,
  name = NULL,
  names = NULL,
  nearest_max_count = NULL,
  period = NULL,
  performance_categories = NULL,
  qualities = NULL,
  reference_time = NULL,
  sensor_levels = NULL,
  ship_code = NULL,
  sources = NULL,
  station_holder = NULL,
  statuses = NULL,
  time_offsets = NULL,
  time_resolutions = NULL,
  time_series_ids = NULL,
  types = NULL,
  unit = NULL,
  units = NULL,
  valid_time = NULL,
  wigos_id = NULL,
  wmo_id = NULL
)
}
\arguments{
\item{endpoint}{Endpoint}

\item{client}{list: List with client id and secret.}

\item{format}{character: Response format.}

\item{version}{character: API version.}

\item{calculation_method}{character: Calculation method}

\item{categories}{character: Categories}

\item{cf_standard_names}{character: CF standard names}

\item{cf_cell_methods}{character: CF cell methods}

\item{cf_units}{character: CF units}

\item{cf_versions}{character: CF versions}

\item{code_tables}{character: Code tables}

\item{country}{character: Country. If specified, only sources whose 'country'
or 'countryCode' attribute matches the search filter may be included in the
result.}

\item{county}{character: County. If specified, only sources whose 'county' or
'countyId' attribute matches the search filter may be included in the
result.}

\item{descriptions}{character: Descriptions}

\item{durations}{character: IDF duration(s), in minutes.}

\item{elements}{character: Elements. See `get_elements()` for available values.}

\item{exposure_categories}{integer: The exposure categories to get
observations for, e.g. 1, 2 etc.}

\item{external_ids}{character: External id. If specified, only sources whose
'externalIds' attribute contains at least one value that matches any of the
specified, comma-separated list of search filters may be included in the
result.}

\item{fields}{character: Fields to return in the response.}

\item{frequencies}{character: IDF frequencies, in years.}

\item{geometry}{character: Sources defined by a specified geometry.
Geometries are specified as either nearest(POINT(...)) or POLYGON(...)
using well-known text representation for geometry (WKT).}

\item{ids}{character: Source ID(s).}

\item{icao_code}{character: ICAO code.}

\item{include_extra}{integer: If 1 extra data is returned.}

\item{lang}{character: ISO language/locale to be used for search filters and
return values}

\item{levels}{numeric: The sensor levels to get observations for, e.g. 0.1,
2, 10, 20.}

\item{level_units}{character: Sensor level units}

\item{level_types}{character: Sensor level types}

\item{limit}{integer: The maximum number of observation times to be returned
for each source/element combination}

\item{location}{character: Geographic position from which to get IDF data, in the
following format "POINT(<longitude degrees> <latitude degrees>)".}

\item{max_age}{character: The maximum observation age as an ISO-8601 period.}

\item{municipality}{character: Municipality. If specified, only sources whose
'municipality' or 'municipalityId' attribute matches the search filter may
be included in the result.}

\item{name}{character: If specified, only sources whose 'name' attribute
matches the search filter be included in the result.}

\item{names}{character: Location name(s)}

\item{nearest_max_count}{character: A string with the maximum number of
sources returned when using "nearest(POINT(...))" for `geometry`. Defaults
to 1.}

\item{period}{character: The validity period, e.g. '1931/1960'.}

\item{performance_categories}{character: The performance categories to get
observations, e.g "A", "B" etc.}

\item{qualities}{integer: The qualities to get observations for, e.g. 1, 2
etc.}

\item{reference_time}{character: Time range to get observations for, in
ISO-8601 format.}

\item{sensor_levels}{character: Sensor levels}

\item{ship_code}{character: Ship code. If specified, only sources whose
'shipCodes' attribute contains at least one name that matches this search
filter may be included in the result.}

\item{sources}{character: Source (station) ID, e.g. 'SN18700'.}

\item{station_holder}{character: Station holder. If specified, only sources
whose 'stationHolders' attribute contains at least one name that matches
the search filter may be included in the result.}

\item{statuses}{character: Statuses}

\item{time_offsets}{character: The time offsets to get observations for,
specified as ISO-8601 periods.}

\item{time_resolutions}{character: The time resolutions to get observations
for, specified as ISO-8601 periods.}

\item{time_series_ids}{integer: The internal time series IDs to get
observations for. Defaults to 0.}

\item{types}{character: Source type. Either "SensorSystem",
"InterpolatedDataset", or "RegionDataset".}

\item{unit}{character: Unit of measure for the intensity, either 'mm' or 'lsha'.}

\item{units}{character: Units}

\item{valid_time}{character: If specified, only sources that have been, or
still are, valid/applicable during some part of this interval may be
included in the result.pecify <date>/<date>, <date>/now, <date>, or now,
where <date> is of the form YYYY-MM-DD, e.g. 2017-03-06. The default is
'now', i.e. only currently valid/applicable sources are included.}

\item{wigos_id}{character: WIGOS ID. If specified, only sources whose
'wigosId' attribute matches the search filter may be included in the
result.}

\item{wmo_id}{character: WMO ID. If specified, only sources whose 'wmoId'
attribute matches the search filter  may be included in the result.}
}
\value{
httr2_request
}
\description{
Create query
}
\keyword{internal}
