% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{create_token}
\alias{create_token}
\title{Create token}
\usage{
create_token(client, return_response = FALSE)
}
\arguments{
\item{client}{list: List with client id and secret. Defaults to the
`MET_FROST_ID` and `MET_FROST_SECRET` environment variables.}

\item{return_response}{logical: If TRUE a list of class `frost_api` is
returned, including the raw `httr2_response`.}
}
\value{
list
}
\description{
Create a FROST API token.
}
\examples{
\dontrun{
# Generate token
create_token()
}
}
\keyword{internal}
