% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climate_normals.R
\name{get_climate_normals}
\alias{get_climate_normals}
\title{Get climate normals}
\usage{
get_climate_normals(
  sources,
  elements = NULL,
  period = NULL,
  format = "jsonld",
  version = "v0",
  client = get_frost_client(),
  auth_type = c("basic", "oauth"),
  flatten = TRUE,
  return_response = FALSE
)
}
\arguments{
\item{sources}{character: Source (station) ID, e.g. 'SN18700'.}

\item{elements}{character: Elements. See `get_elements()` for available
values.}

\item{period}{character: The validity period, e.g. '1931/1960'.}

\item{format}{character: Response format.}

\item{version}{character: API version.}

\item{client}{list: List with client id and secret. Defaults to the
`MET_FROST_ID` and `MET_FROST_SECRET` environment variables.}

\item{auth_type}{character: Authentication method, either 'basic' or 'oauth'.}

\item{flatten}{logical: If TRUE the response is transformed to a table.}

\item{return_response}{logical: If TRUE a list of class `frost_api` is
returned, including the raw `httr2_response`.}
}
\description{
Get climate normals.
}
\examples{
df <- get_climate_normals("SN100")
}
