% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elements.R
\name{get_code_tables}
\alias{get_code_tables}
\title{Get code tables}
\usage{
get_code_tables(
  ids = NULL,
  fields = NULL,
  lang = c("en-US", "nb-NO", "nn-NO"),
  version = "v0",
  format = "jsonld",
  client = get_frost_client(),
  auth_type = c("basic", "oauth"),
  flatten = TRUE,
  return_response = FALSE
)
}
\arguments{
\item{ids}{character: Source ID(s).}

\item{fields}{character: Fields to return in the response.}

\item{lang}{character: ISO language/locale to be used for search filters and
return values}

\item{version}{character: API version.}

\item{format}{character: Response format.}

\item{client}{list: List with client id and secret. Defaults to the
`MET_FROST_ID` and `MET_FROST_SECRET` environment variables.}

\item{auth_type}{character: Authentication method, either 'basic' or 'oauth'.}

\item{flatten}{logical: If TRUE the response is transformed to a table.}

\item{return_response}{logical: If TRUE a list of class `frost_api` is
returned, including the raw `httr2_response`.}
}
\description{
Get metadata about the code tables.
}
\details{
Use the default query parameters to retrieve all code tables.
}
\examples{
\dontrun{
# Get all code tables
df <- get_code_tables()
}
}
