% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_records.R
\name{get_county_extremes}
\alias{get_county_extremes}
\title{Get county extremes}
\usage{
get_county_extremes(
  sources = NULL,
  sourcenames = NULL,
  counties = NULL,
  municipalities = NULL,
  elements = NULL,
  months = NULL,
  fields = NULL,
  version = "v0",
  format = "jsonld",
  client = get_frost_client(),
  auth_type = c("basic", "oauth"),
  simplify = TRUE,
  return_response = FALSE
)
}
\arguments{
\item{sources}{character: Source (station) ID, e.g. 'SN18700'.}

\item{sourcenames}{character: Source names.}

\item{counties}{character: Counties}

\item{municipalities}{character: Municipalities}

\item{elements}{character: Elements. See `get_elements()` for available
values.}

\item{months}{character: Months to get records for as integers or integer
ranges between 1 and 12, e.g. '1,5,8-12'.}

\item{fields}{character: Fields to return in the response.}

\item{version}{character: API version.}

\item{format}{character: Response format.}

\item{client}{list: List with client id and secret. Defaults to the
`MET_FROST_ID` and `MET_FROST_SECRET` environment variables.}

\item{auth_type}{character: Authentication method, either 'basic' or 'oauth'.}

\item{simplify}{logical: If TRUE the response is transformed to a standardized table.}

\item{return_response}{logical: If TRUE a list of class `frost_api` is
returned, including the raw `httr2_response`.}
}
\value{
tibble or list
}
\description{
Use the default query parameters to retrieve all values.
}
\examples{
\dontrun{
# Get all sources
df <- get_county_extremes()
}
}
