% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elements.R
\name{get_elements}
\alias{get_elements}
\title{Get elements}
\usage{
get_elements(
  ids = NULL,
  names = NULL,
  descriptions = NULL,
  units = NULL,
  code_tables = NULL,
  statuses = NULL,
  calculation_method = NULL,
  categories = NULL,
  time_offsets = NULL,
  sensor_levels = NULL,
  cf_standard_names = NULL,
  cf_cell_methods = NULL,
  cf_units = NULL,
  cf_versions = NULL,
  fields = NULL,
  lang = c("en-US", "nb-NO", "nn-NO"),
  version = "v0",
  format = c("jsonld"),
  client = get_api_client(),
  flatten = TRUE,
  return_response = FALSE
)
}
\arguments{
\item{ids}{character: Source ID(s).}

\item{names}{character: Location name(s)}

\item{descriptions}{character: Descriptions}

\item{units}{character: Units}

\item{code_tables}{character: Code tables}

\item{statuses}{character: Statuses}

\item{calculation_method}{character: Calculation method}

\item{categories}{character: Categories}

\item{time_offsets}{character: The time offsets to get observations for,
specified as ISO-8601 periods.}

\item{sensor_levels}{character: Sensor levels}

\item{cf_standard_names}{character: CF standard names}

\item{cf_cell_methods}{character: CF cell methods}

\item{cf_units}{character: CF units}

\item{cf_versions}{character: CF versions}

\item{fields}{character: Fields to return in the response.}

\item{lang}{character: ISO language/locale to be used for search filters and
return values}

\item{version}{character: API version.}

\item{format}{character: Response format.}

\item{client}{list: List with client id and secret.}

\item{flatten}{logical: If TRUE the response is transformed to an unnested
table.}

\item{return_response}{logical: If TRUE a list of class `frost_api` is
returned, including the raw `httr2_response`.}
}
\value{
tibble or list
}
\description{
Get metadata about the weather and climate elements.
}
\details{
Use the default query parameters to retrieve all elements.
}
\examples{
\dontrun{
# Get all elements
df <- get_elements()
}
}
