% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_frost_client}
\alias{get_frost_client}
\title{Get Frost API Client}
\usage{
get_frost_client()
}
\description{
Get Frost API client info from R environment variables.
}
\examples{
\dontrun{
get_frost_client()
}
}
