% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations.R
\name{get_locations}
\alias{get_locations}
\title{Get locations}
\usage{
get_locations(
  names = NULL,
  geometry = NULL,
  fields = NULL,
  version = "v0",
  format = "jsonld",
  client = get_api_client(),
  flatten = TRUE,
  return_response = FALSE
)
}
\arguments{
\item{names}{character: Location name(s)}

\item{geometry}{character: Sources defined by a specified geometry.
Geometries are specified as either nearest(POINT(...)) or POLYGON(...)
using well-known text representation for geometry (WKT).}

\item{fields}{character: Fields to return in the response.}

\item{version}{character: API version.}

\item{format}{character: Response format.}

\item{client}{list: List with client id and secret.}

\item{flatten}{logical: If TRUE the response is transformed to an unnested
table.}

\item{return_response}{logical: If TRUE a list of class `frost_api` is
returned, including the raw `httr2_response`.}
}
\description{
Get metadata for location names.
}
\details{
Use the default query parameters to retrieve all location names.
}
\examples{
\dontrun{
df <- get_locations()
}
}
