% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations.R
\name{get_observations_ts}
\alias{get_observations_ts}
\title{Get observations time series}
\usage{
get_observations_ts(
  sources = NULL,
  reference_time = NULL,
  elements = NULL,
  time_offsets = NULL,
  time_resolutions = NULL,
  time_series_ids = NULL,
  performance_categories = NULL,
  exposure_categories = NULL,
  qualities = NULL,
  levels = NULL,
  level_types = NULL,
  level_units = NULL,
  include_extra = NULL,
  fields = NULL,
  version = "v0",
  format = "jsonld",
  client = get_frost_client(),
  simplify = TRUE,
  return_response = FALSE
)
}
\arguments{
\item{sources}{character: Source (station) ID, e.g. 'SN18700'.}

\item{reference_time}{character: Time range to get observations for, in
ISO-8601 format.}

\item{elements}{character: Elements. See `get_elements()` for available
values.}

\item{time_offsets}{character: The time offsets to get observations for,
specified as ISO-8601 periods.}

\item{time_resolutions}{character: The time resolutions to get observations
for, specified as ISO-8601 periods.}

\item{time_series_ids}{integer: The internal time series IDs to get
observations for. Defaults to 0.}

\item{performance_categories}{character: The performance categories to get
observations, e.g "A", "B" etc.}

\item{exposure_categories}{integer: The exposure categories to get
observations for, e.g. 1, 2 etc.}

\item{qualities}{integer: The qualities to get observations for, e.g. 1, 2
etc.}

\item{levels}{numeric: The sensor levels to get observations for, e.g. 0.1,
2, 10, 20.}

\item{level_types}{character: Sensor level types}

\item{level_units}{character: Sensor level units}

\item{include_extra}{integer: If 1 extra data is returned.}

\item{fields}{character: Fields to return in the response.}

\item{version}{character: API version.}

\item{format}{character: Response format.}

\item{client}{list: List with client id and secret. Defaults to the
`MET_FROST_ID` and `MET_FROST_SECRET` environment variables.}

\item{simplify}{logical: If TRUE the response is transformed to a standardized table.}

\item{return_response}{logical: If TRUE a list of class `frost_api` is
returned, including the raw `httr2_response`.}
}
\description{
Get available time series for observations.
}
\details{
Use the default query parameters to retrieve all time series.
}
\examples{
\dontrun{
df <- get_observations_ts("SN23670")
}
}
